Moved state/exceptions.py to utilities
