+++ src/mepo/repository/git.py
@@ -8,7 +8,7 @@ from urllib.parse import urljoin
 from ..utilities import shellcmd
 from ..utilities import colors
 
-from ..state.exceptions import RepoAlreadyClonedError
+from ..utilities.exceptions import RepoAlreadyClonedError
 
 class GitRepository(object):
     """
